# health-portal/src/main/java/com/mycompany/models/ClinicListResponse.java
package com.mycompany.models;

import java.util.List;
import java.util.Map;

import com.fasterxml.jackson.annotation.JsonProperty;

import lombok.AllArgsConstructor;
import lombok.Getter;

@Getter
@AllArgsConstructor
public class ClinicListResponse {
    private List<Clinic> content;
    private PageMetadata page;
    @JsonProperty("_links")
    private Map<String, Link> links;
}