# health-portal/src/main/java/com/mycompany/models/Link.java
package com.mycompany.models;

import lombok.AllArgsConstructor;
import lombok.Getter;

@Getter
@AllArgsConstructor
public class Link {
    private String href;
    private String rel;
    private String method;
}